Revert "Evaluation component"
