+++ src/components/tools/pairwise-comparison/steps/PCPairComparison/PCPairComparisonComponent.tsx
@@ -3,12 +3,10 @@ import {
     Step,
     StepProp
 } from "../../../../../general-components/Tool/SteppableTool/StepComponent/Step/Step";
+import {CompareComponent, CompareValue} from "../../../../../general-components/CompareComponent/CompareComponent";
 import {
-    CompareComponent,
-    CompareComponentValues,
-    CompareValue
-} from "../../../../../general-components/CompareComponent/CompareComponent";
-import {MatchCardComponentFieldsAdapter} from "../../../../../general-components/CompareComponent/Adapter/MatchCardComponentFieldsAdapter";
+    MatchCardComponentFieldsAdapter
+} from "../../../../../general-components/CompareComponent/Adapter/MatchCardComponentFieldsAdapter";
 import {PairwiseComparisonValues} from "../../PairwiseComparison";
 import {PCPairComparison} from "./PCPairComparison";
 import {CompareHeader} from "../../../../../general-components/CompareComponent/Header/CompareHeaderAdapter";
@@ -17,7 +15,8 @@ import {CompareHeader} from "../../../../../general-components/CompareComponent/
 /**
  * Die Werte des zweiten Schrittes des Paarweisen-Vergleiches
  */
-export interface PCPairComparisonValues extends CompareComponentValues {
+export interface PCPairComparisonValues {
+    comparisons: CompareValue[],
     headers: CompareHeader[]
 }
 
@@ -29,6 +28,7 @@ class PCPairComparisonComponent extends Step<PairwiseComparisonValues, {}> {
     public constructor(props: Readonly<StepProp<PairwiseComparisonValues>> | StepProp<PairwiseComparisonValues>, context: any) {
         super(props, context);
     }
+
 
     shouldComponentUpdate(nextProps: Readonly<StepProp<PairwiseComparisonValues>>, nextState: Readonly<{}>, nextContext: any): boolean {
         let shouldUpdate = !shallowCompareStepProps(this.props, nextProps);
@@ -52,6 +52,7 @@ class PCPairComparisonComponent extends Step<PairwiseComparisonValues, {}> {
         const criteriaValues = this.props.save.data["pc-criterias"]?.criterias;
 
         if (comparisonValues !== undefined && criteriaValues !== undefined) {
+
             // TODO remove adapter
             const adapter = new MatchCardComponentFieldsAdapter(criteriaValues);
