+++ src/general-components/CompareComponent/CompareComponent.tsx
@@ -1,6 +1,6 @@
 import {Component} from "react";
 import {CompareAdapter} from "./Adapter/CompareAdapter";
-import {CompareHeader, CompareHeaderAdapter} from "./Header/CompareHeaderAdapter";
+import {CompareHeaderAdapter} from "./Header/CompareHeaderAdapter";
 
 import "./compare-component.scss";
 
@@ -12,9 +12,6 @@ type CompareValue = {
     value: null | string,
     header: null | string,
 };
-type CompareComponentValues = {
-    comparisons: CompareValue[]
-}
 
 export interface CompareComponentProps {
     /**
@@ -157,6 +154,5 @@ export {
 };
 
 export type{
-    CompareValue,
-    CompareComponentValues
+    CompareValue
 }
